# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pllay On Edge Athlete Edition API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /athlete/profile:
    get:
      operationId: getAthleteProfile
      tags: [athlete]
      summary: Get the current athlete's profile
      responses:
        "200":
          description: Athlete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfile"
        "404":
          description: Profile not found (not yet created)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createAthleteProfile
      tags: [athlete]
      summary: Create or update the athlete's profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AthleteProfileInput"
      responses:
        "200":
          description: Profile created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteProfile"
...
  /phase0/modules/{moduleName}:
    get:
      operationId: getPhase0Module
      tags: [phase0]
      summary: Get a specific Phase 0 module
      parameters:
        - name: moduleName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Phase 0 module data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase0Module"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: savePhase0Module
      tags: [phase0]
      summary: Save a Phase 0 module (upsert)
      parameters:
        - name: moduleName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Phase0ModuleInput"
      responses:
        "200":
          description: Module saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase0Module"
...
  /competition-reviews:
    get:
      operationId: getCompetitionReviews
      tags: [competition]
      summary: Get all competition reviews for the current athlete
      responses:
        "200":
          description: List of competition reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompetitionReview"
    post:
      operationId: createCompetitionReview
      tags: [competition]
      summary: Create a new competition review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionReview"
  /competition-reviews/{id}:
    put:
      operationId: updateCompetitionReview
      tags: [competition]
      summary: Update a competition review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionReviewInput"
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionReview"
  /monthly-checkins/{monthNumber}:
    get:
      operationId: getMonthlyCheckin
      tags: [checkin]
      summary: Get a monthly check-in
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly check-in data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyCheckin"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: saveMonthlyCheckin
      tags: [checkin]
      summary: Save a monthly check-in (upsert)
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonthlyCheckinInput"
      responses:
        "200":
          description: Check-in saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyCheckin"
  /capstone:
    get:
      operationId: getCapstone
      tags: [capstone]
      summary: Get the capstone review
      responses:
        "200":
          description: Capstone data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capstone"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: saveCapstone
      tags: [capstone]
      summary: Save the capstone review (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapstoneInput"
      responses:
        "200":
          description: Capstone saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capstone"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AthleteProfile:
      type: object
      required: [id, userId, name, dob, sport, programStartDate]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        dob:
          type: string
        sport:
          type: string
        club:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        parentName:
          type: ["string", "null"]
        programStartDate:
          type: string
        phase0CompletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AthleteProfileInput:
      type: object
      required: [name, dob, sport]
      properties:
        name:
          type: string
        dob:
          type: string
        sport:
          type: string
        club:
          type: string
        coachName:
          type: string
        parentName:
          type: string
    Phase0Module:
      type: object
      required: [id, athleteId, moduleName, completed]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        moduleName:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        data:
          type: object
    Phase0ModuleInput:
      type: object
      required: [data, completed]
      properties:
        data:
          type: object
        completed:
          type: boolean
...
    CompetitionReview:
      type: object
      required: [id, athleteId]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        phase:
          type: ["string", "null"]
        competitionName:
          type: ["string", "null"]
        competitionDate:
          type: ["string", "null"]
        opponent:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        performanceRating:
          type: ["integer", "null"]
        decisionRating:
          type: ["integer", "null"]
        emotionRating:
          type: ["integer", "null"]
        bestDecision:
          type: ["string", "null"]
        changeDecision:
          type: ["string", "null"]
        keyLearning:
          type: ["string", "null"]
        implWhen:
          type: ["string", "null"]
        implWhere:
          type: ["string", "null"]
        implHow:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
    CompetitionReviewInput:
      type: object
      properties:
        phase:
          type: string
        competitionName:
          type: string
        competitionDate:
          type: string
        opponent:
          type: string
        result:
          type: string
        performanceRating:
          type: integer
        decisionRating:
          type: integer
        emotionRating:
          type: integer
        bestDecision:
          type: string
        changeDecision:
          type: string
        keyLearning:
          type: string
        implWhen:
          type: string
        implWhere:
          type: string
        implHow:
          type: string
    MonthlyCheckin:
      type: object
      required: [id, athleteId, monthNumber]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        monthNumber:
          type: integer
        completedAt:
          type: ["string", "null"]
        effortAvg:
          type: ["number", "null"]
        focusAvg:
          type: ["number", "null"]
        consistencyAvg:
          type: ["number", "null"]
        recoveryAvg:
          type: ["number", "null"]
        ownershipAvg:
          type: ["number", "null"]
        goal1Progress:
          type: ["string", "null"]
        goal2Progress:
          type: ["string", "null"]
        behaviour1Status:
          type: ["string", "null"]
        keyHabit:
          type: ["string", "null"]
        keyLearning:
          type: ["string", "null"]
        changeNextPhase:
          type: ["string", "null"]
        data:
          type: object
    MonthlyCheckinInput:
      type: object
      properties:
        effortAvg:
          type: number
        focusAvg:
          type: number
        consistencyAvg:
          type: number
        recoveryAvg:
          type: number
        ownershipAvg:
          type: number
        goal1Progress:
          type: string
        goal2Progress:
          type: string
        behaviour1Status:
          type: string
        keyHabit:
          type: string
        keyLearning:
          type: string
        changeNextPhase:
          type: string
        data:
          type: object
        completed:
          type: boolean
    Capstone:
      type: object
      required: [id, athleteId]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        completedAt:
          type: ["string", "null"]
        physicalStart:
          type: ["integer", "null"]
        physicalNow:
          type: ["integer", "null"]
        technicalStart:
          type: ["integer", "null"]
        technicalNow:
          type: ["integer", "null"]
        tacticalStart:
          type: ["integer", "null"]
        tacticalNow:
          type: ["integer", "null"]
        psychologicalStart:
          type: ["integer", "null"]
        psychologicalNow:
          type: ["integer", "null"]
        lifestyleStart:
          type: ["integer", "null"]
        lifestyleNow:
          type: ["integer", "null"]
        goal1Achieved:
          type: ["string", "null"]
        goal2Achieved:
          type: ["string", "null"]
        goal3Achieved:
          type: ["string", "null"]
        evidenceCompetition:
          type: ["string", "null"]
        evidenceTechnical:
          type: ["string", "null"]
        evidenceBehaviour:
          type: ["string", "null"]
        fieldTest1:
          type: ["boolean", "null"]
        fieldTest2:
          type: ["boolean", "null"]
        fieldTestNotes:
          type: ["string", "null"]
        engagementRating:
          type: ["integer", "null"]
        awarenessRating:
          type: ["integer", "null"]
        decisionsRating:
          type: ["integer", "null"]
        accountabilityRating:
          type: ["integer", "null"]
        proudMoment:
          type: ["string", "null"]
        fellShort:
          type: ["string", "null"]
        learned:
          type: ["string", "null"]
        doDifferently:
          type: ["string", "null"]
        signatureName:
          type: ["string", "null"]
    CapstoneInput:
      type: object
      properties:
        physicalStart:
          type: integer
        physicalNow:
          type: integer
        technicalStart:
          type: integer
        technicalNow:
          type: integer
        tacticalStart:
          type: integer
        tacticalNow:
          type: integer
        psychologicalStart:
          type: integer
        psychologicalNow:
          type: integer
        lifestyleStart:
          type: integer
        lifestyleNow:
          type: integer
        goal1Achieved:
          type: string
        goal2Achieved:
          type: string
        goal3Achieved:
          type: string
        evidenceCompetition:
          type: string
        evidenceTechnical:
          type: string
        evidenceBehaviour:
          type: string
        fieldTest1:
          type: boolean
        fieldTest2:
          type: boolean
        fieldTestNotes:
          type: string
        engagementRating:
          type: integer
        awarenessRating:
          type: integer
        decisionsRating:
          type: integer
        accountabilityRating:
          type: integer
        proudMoment:
          type: string
        fellShort:
          type: string
        learned:
          type: string
        doDifferently:
          type: string
        signatureName:
          type: string
        completed:
          type: boolean
...