# lib/api-zod/src/generated/types/capstone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */

export interface Capstone {
  id: number;
  athleteId: number;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  physicalStart?: number | null;
  /** @nullable */
  physicalNow?: number | null;
  /** @nullable */
  technicalStart?: number | null;
  /** @nullable */
  technicalNow?: number | null;
  /** @nullable */
  tacticalStart?: number | null;
  /** @nullable */
  tacticalNow?: number | null;
  /** @nullable */
  psychologicalStart?: number | null;
  /** @nullable */
  psychologicalNow?: number | null;
  /** @nullable */
  lifestyleStart?: number | null;
  /** @nullable */
  lifestyleNow?: number | null;
  /** @nullable */
  goal1Achieved?: string | null;
  /** @nullable */
  goal2Achieved?: string | null;
  /** @nullable */
  goal3Achieved?: string | null;
  /** @nullable */
  evidenceCompetition?: string | null;
  /** @nullable */
  evidenceTechnical?: string | null;
  /** @nullable */
  evidenceBehaviour?: string | null;
  /** @nullable */
  fieldTest1?: boolean | null;
  /** @nullable */
  fieldTest2?: boolean | null;
  /** @nullable */
  fieldTestNotes?: string | null;
  /** @nullable */
  engagementRating?: number | null;
  /** @nullable */
  awarenessRating?: number | null;
  /** @nullable */
  decisionsRating?: number | null;
  /** @nullable */
  accountabilityRating?: number | null;
  /** @nullable */
  proudMoment?: string | null;
  /** @nullable */
  fellShort?: string | null;
  /** @nullable */
  learned?: string | null;
  /** @nullable */
  doDifferently?: string | null;
  /** @nullable */
  signatureName?: string | null;
}