# lib/api-zod/src/generated/types/capstoneInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */

export interface CapstoneInput {
  physicalStart?: number;
  physicalNow?: number;
  technicalStart?: number;
  technicalNow?: number;
  tacticalStart?: number;
  tacticalNow?: number;
  psychologicalStart?: number;
  psychologicalNow?: number;
  lifestyleStart?: number;
  lifestyleNow?: number;
  goal1Achieved?: string;
  goal2Achieved?: string;
  goal3Achieved?: string;
  evidenceCompetition?: string;
  evidenceTechnical?: string;
  evidenceBehaviour?: string;
  fieldTest1?: boolean;
  fieldTest2?: boolean;
  fieldTestNotes?: string;
  engagementRating?: number;
  awarenessRating?: number;
  decisionsRating?: number;
  accountabilityRating?: number;
  proudMoment?: string;
  fellShort?: string;
  learned?: string;
  doDifferently?: string;
  signatureName?: string;
  completed?: boolean;
}