# lib/api-zod/src/generated/types/monthlyCheckin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyCheckinData } from './monthlyCheckinData';

export interface MonthlyCheckin {
  id: number;
  athleteId: number;
  monthNumber: number;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  effortAvg?: number | null;
  /** @nullable */
  focusAvg?: number | null;
  /** @nullable */
  consistencyAvg?: number | null;
  /** @nullable */
  recoveryAvg?: number | null;
  /** @nullable */
  ownershipAvg?: number | null;
  /** @nullable */
  goal1Progress?: string | null;
  /** @nullable */
  goal2Progress?: string | null;
  /** @nullable */
  behaviour1Status?: string | null;
  /** @nullable */
  keyHabit?: string | null;
  /** @nullable */
  keyLearning?: string | null;
  /** @nullable */
  changeNextPhase?: string | null;
  data?: MonthlyCheckinData;
}