# lib/api-zod/src/generated/types/monthlyCheckinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyCheckinInputData } from './monthlyCheckinInputData';

export interface MonthlyCheckinInput {
  effortAvg?: number;
  focusAvg?: number;
  consistencyAvg?: number;
  recoveryAvg?: number;
  ownershipAvg?: number;
  goal1Progress?: string;
  goal2Progress?: string;
  behaviour1Status?: string;
  keyHabit?: string;
  keyLearning?: string;
  changeNextPhase?: string;
  data?: MonthlyCheckinInputData;
  completed?: boolean;
}