# lib/api-zod/src/generated/types/phase0Module.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */
import type { Phase0ModuleData } from './phase0ModuleData';

export interface Phase0Module {
  id: number;
  athleteId: number;
  moduleName: string;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  data?: Phase0ModuleData;
}